datingTestSet
DatingTestSet.txt is Helen's dating data. The attributes which Helen cares about is the annual frequent flier mileage, video game time percentage, liters of ice-cream consuming per week, and the samples are labeled as 'in large doses', 'in small doses', and 'not at all'. Try to improve the process of choosing a date for Helen.